Refactor SnippetCodeGraph to constructor initialization
