+++ CodeSnippetsReflection.OpenAPI.Test/AssertUtil.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace CodeSnippetsReflection.OpenAPI.Test
+{
+    public static class AssertUtil
+    {
+        public static void ContainsIgnoreWhiteSpace(string expectedSubstring, string actualString)
+        {
+            Xunit.Assert.Contains(
+                expectedSubstring.Replace(" ", string.Empty).Replace(Environment.NewLine, string.Empty),
+                actualString.Replace(" ", string.Empty).Replace(Environment.NewLine, string.Empty)
+            );
+        }
+    }
+}
