+++ CodeSnippetsReflection.OpenAPI/CodeSnippetsReflection.OpenAPI.csproj
@@ -8,7 +8,6 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.ApplicationInsights" Version="2.20.0" />
     <PackageReference Include="Microsoft.OpenApi.Readers" Version="1.3.1-preview5" />
-    <PackageReference Include="System.Linq" Version="4.3.0" />
     <PackageReference Include="System.Text.Json" Version="6.0.4" />
   </ItemGroup>
